+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/composer/MessageComposer.kt
@@ -37,7 +37,7 @@ import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.common.state.Edit
 import io.getstream.chat.android.common.state.ValidationError
 import io.getstream.chat.android.compose.R
-import io.getstream.chat.android.compose.state.messages.composer.MessageInputState
+import io.getstream.chat.android.compose.state.messages.composer.MessageComposerState
 import io.getstream.chat.android.compose.ui.messages.composer.components.DefaultComposerIntegrations
 import io.getstream.chat.android.compose.ui.messages.composer.components.MessageInput
 import io.getstream.chat.android.compose.ui.messages.composer.components.MessageInputOptions
@@ -63,6 +63,7 @@ import io.getstream.chat.android.compose.viewmodel.messages.MessageComposerViewM
  * @param onMentionSelected Handler when the user taps on a mention suggestion item.
  * @param onCommandSelected Handler when the user taps on a command suggestion item.
  * @param mentionPopupContent Customizable composable function that represents the mention suggestions popup.
+ * @param commandPopupContent Customizable composable function that represents the instant command suggestions popup.
  * @param integrations A view that represents custom integrations. By default, we provide
  * [DefaultComposerIntegrations], which show Attachments & Giphy, but users can override this with
  * their own integrations, which they need to hook up to their own data providers and UI.
@@ -93,7 +94,7 @@ public fun MessageComposer(
             onCommandSelected = onCommandSelected
         )
     },
-    integrations: @Composable RowScope.(MessageInputState) -> Unit = {
+    integrations: @Composable RowScope.(MessageComposerState) -> Unit = {
         DefaultComposerIntegrations(
             modifier = Modifier
                 .height(44.dp)
@@ -104,14 +105,14 @@ public fun MessageComposer(
         )
     },
     label: @Composable () -> Unit = { DefaultComposerLabel() },
-    input: @Composable RowScope.(MessageInputState) -> Unit = {
+    input: @Composable RowScope.(MessageComposerState) -> Unit = {
         MessageInput(
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(vertical = 8.dp)
                 .weight(1f),
             label = label,
-            messageInputState = it,
+            messageComposerState = it,
             onValueChange = onValueChange,
             onAttachmentRemoved = onAttachmentRemoved
         )
@@ -138,7 +139,7 @@ public fun MessageComposer(
         commandPopupContent = commandPopupContent,
         integrations = integrations,
         input = input,
-        messageInputState = MessageInputState(
+        messageComposerState = MessageComposerState(
             inputValue = value,
             attachments = selectedAttachments,
             action = activeAction,
@@ -156,24 +157,26 @@ public fun MessageComposer(
  * Clean version of the [MessageComposer] that doesn't rely on ViewModels, so the user can provide a
  * manual way to handle and represent data and various operations.
  *
- * @param messageInputState The state of the message input.
+ * @param messageComposerState The state of the message input.
  * @param onSendMessage Handler when the user taps on the send message button.
  * @param onCancelAction Handler when the user cancels the active action (Reply or Edit).
  * @param onMentionSelected Handler when the user taps on a mention suggestion item.
  * @param onCommandSelected Handler when the user taps on a command suggestion item.
- * @param mentionPopupContent Customizable composable function that represents the mention suggestions popup.
  * @param modifier Modifier for styling.
+ * @param mentionPopupContent Customizable composable function that represents the mention suggestions popup.
+ * @param commandPopupContent Customizable composable function that represents the instant command suggestions popup.
  * @param shouldShowIntegrations If we should show or hide integrations.
  * @param integrations Composable that represents integrations for the composer, such as Attachments.
  * @param input Composable that represents the input field in the composer.
  */
 @Composable
 public fun MessageComposer(
-    messageInputState: MessageInputState,
+    messageComposerState: MessageComposerState,
     onSendMessage: (String, List<Attachment>) -> Unit,
     onCancelAction: () -> Unit,
     onMentionSelected: (User) -> Unit,
     onCommandSelected: (Command) -> Unit,
+    modifier: Modifier = Modifier,
     mentionPopupContent: @Composable (List<User>) -> Unit = {
         DefaultMentionPopupContent(
             mentionSuggestions = it,
@@ -186,14 +189,13 @@ public fun MessageComposer(
             onCommandSelected = onCommandSelected
         )
     },
-    modifier: Modifier = Modifier,
     shouldShowIntegrations: Boolean = true,
-    integrations: @Composable RowScope.(MessageInputState) -> Unit,
-    input: @Composable RowScope.(MessageInputState) -> Unit,
+    integrations: @Composable RowScope.(MessageComposerState) -> Unit,
+    input: @Composable RowScope.(MessageComposerState) -> Unit,
 ) {
-    val (value, attachments, activeAction, validationErrors, mentionSuggestions, commandSuggestions, cooldownTimer) = messageInputState
+    val (value, attachments, activeAction, validationErrors, mentionSuggestions, commandSuggestions, cooldownTimer) = messageComposerState
 
-    showValidationErrorIfNecessary(validationErrors)
+    MessageInputValidationError(validationErrors)
 
     Surface(
         modifier = modifier,
@@ -220,14 +222,14 @@ public fun MessageComposer(
             ) {
 
                 if (shouldShowIntegrations && activeAction !is Edit) {
-                    integrations(messageInputState)
+                    integrations(messageComposerState)
                 } else {
                     Spacer(
                         modifier = Modifier.size(16.dp)
                     )
                 }
 
-                input(messageInputState)
+                input(messageComposerState)
 
                 if (cooldownTimer > 0) {
                     CooldownIndicator(cooldownTimer = cooldownTimer)
@@ -311,7 +313,7 @@ internal fun DefaultMentionPopupContent(
  * Represents the default command suggestion list popup shown above the message composer.
  *
  * @param commandSuggestions The list of available commands in the channel.
- * @param onCommandClick Handler when the user taps on a command suggestion item.
+ * @param onCommandSelected Handler when the user taps on a command suggestion item.
  */
 @Composable
 internal fun DefaultCommandPopupContent(
@@ -345,7 +347,7 @@ internal fun DefaultComposerLabel() {
  * @param validationErrors The list of validation errors for the current user input.
  */
 @Composable
-private fun showValidationErrorIfNecessary(validationErrors: List<ValidationError>) {
+private fun MessageInputValidationError(validationErrors: List<ValidationError>) {
     if (validationErrors.isNotEmpty()) {
         val errorMessage = when (val validationError = validationErrors.first()) {
             is ValidationError.MessageLengthExceeded -> {