+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/StreamDimens.kt
@@ -32,6 +32,9 @@ import androidx.compose.ui.unit.dp
  * @param commandSuggestionItemHorizontalPadding The horizontal content padding inside command list item.
  * @param commandSuggestionItemVerticalPadding The vertical content padding inside command list item.
  * @param commandSuggestionItemIconSize The size of a command icon in the suggestion list popup.
+ * @param messageOptionsMaxHeight The max height of the message options section when we select a message in the list.
+ * @param messageOptionsMaxWidth The max width of the message options section when we select a message in the list.
+ * @param messageOptionsRoundedCorners The rounded corners size of the message options shape.
  */
 @Immutable
 public data class StreamDimens(
@@ -60,6 +63,9 @@ public data class StreamDimens(
     public val commandSuggestionItemHorizontalPadding: Dp,
     public val commandSuggestionItemVerticalPadding: Dp,
     public val commandSuggestionItemIconSize: Dp,
+    public val messageOptionsMaxWidth: Dp,
+    public val messageOptionsMaxHeight: Dp,
+    public val messageOptionsRoundedCorners: Dp,
 ) {
     public companion object {
         public fun defaultDimens(): StreamDimens = StreamDimens(
@@ -88,6 +94,9 @@ public data class StreamDimens(
             commandSuggestionItemHorizontalPadding = 8.dp,
             commandSuggestionItemVerticalPadding = 8.dp,
             commandSuggestionItemIconSize = 24.dp,
+            messageOptionsMaxWidth = 200.dp,
+            messageOptionsMaxHeight = 300.dp,
+            messageOptionsRoundedCorners = 16.dp
         )
     }
 }